+++ com.unity.render-pipelines.core/Runtime/Debugging/DebugUI.Fields.cs
@@ -83,7 +83,7 @@ namespace UnityEngine.Rendering
             /// Set the value of the field.
             /// </summary>
             /// <param name="value">Input value.</param>
-            public void SetValue(T value)
+            public virtual void SetValue(T value)
             {
                 Assert.IsNotNull(setter);
                 var v = ValidateValue(value);
@@ -290,7 +290,9 @@ namespace UnityEngine.Rendering
             public int[] enumValues;
 
             internal int[] quickSeparators;
-            internal int[] indexes;
+
+            private int[] m_Indexes;
+            internal int[] indexes => m_Indexes ??= Enumerable.Range(0, enumNames?.Length ?? 0).ToArray();
 
             /// <summary>
             /// Get the enumeration value index.
@@ -304,7 +306,11 @@ namespace UnityEngine.Rendering
             /// <summary>
             /// Current enumeration value index.
             /// </summary>
-            public int currentIndex { get => getIndex(); set => setIndex(value); }
+            public int currentIndex
+            {
+                get => getIndex();
+                set => setIndex(value);
+            }
 
             /// <summary>
             /// Generates enumerator values and names automatically based on the provided type.
@@ -315,7 +321,6 @@ namespace UnityEngine.Rendering
                 {
                     enumNames = EnumUtility.MakeEnumNames(value);
                     enumValues = EnumUtility.MakeEnumValues(value);
-                    InitIndexes();
                     InitQuickSeparators();
                 }
             }
@@ -344,15 +349,25 @@ namespace UnityEngine.Rendering
                 }
             }
 
-            internal void InitIndexes()
+            /// <summary>
+            /// Set the value of the field.
+            /// </summary>
+            /// <param name="value">Input value.</param>
+            public override void SetValue(int value)
             {
-                if (enumNames == null)
-                    enumNames = new GUIContent[0];
+                Assert.IsNotNull(setter);
+                var validValue = ValidateValue(value);
 
-                indexes = new int[enumNames.Length];
-                for (int i = 0; i < enumNames.Length; i++)
+                // There might be cases that the value does not map the index, look for the correct index
+                var newCurrentIndex = Array.IndexOf(enumValues, validValue);
+
+                if (currentIndex != newCurrentIndex && !validValue.Equals(getter()))
                 {
-                    indexes[i] = i;
+                    setter(validValue);
+                    onValueChanged?.Invoke(this, validValue);
+
+                    if (newCurrentIndex > -1)
+                        currentIndex = newCurrentIndex;
                 }
             }
         }