+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDStencilUsage.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -53,11 +54,16 @@ namespace UnityEngine.Rendering.HighDefinition
     ///         Comp Always
     ///         Pass Replace
     /// </summary>
+    [Flags]
     public enum UserStencilUsage
     {
+        /// <summary>All stencil bits disabled.</summary>
+        None = 0,
         /// <summary>User stencil bit 0.</summary>
         UserBit0 = StencilUsage.UserBit0,
         /// <summary>User stencil bit 1.</summary>
-        UserBit1 = StencilUsage.UserBit1
+        UserBit1 = StencilUsage.UserBit1,
+        /// <summary>The mask of all the user bits.</summary>
+        AllUserBits = UserBit0 | UserBit1,
     }
 }