+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -22,11 +22,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added time slicing support for realtime reflection probes.
 - Added denoising for the path tracer.
 - Added an initial version of under water rendering for the water system.
-- Added Asymmetric projection and Screen Coordinates Override frame settings. Adapted post effects to support Screen Coordinates Override. (Used, for example, to support Cluster Display.)
 - Added option to animate APV sample noise to smooth it out when TAA is enabled.
 - Added default DOTS compatible loading shader (MaterialLoading.shader)
 - Add #pragma editor_sync_compilation directive to MaterialError.shader
 - Added the culling matrix and near plane for lights, so that they can be custom-culled with the BatchRenderGroup API.
+- Added an optional CPU simulation for the water system.
+- Added new Unity material ball matching the new Unity logo.
 
 ### Changed
 - Render Graph object pools are now cleared with render graph cleanup
@@ -136,6 +137,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed camera motion vector pass reading last frame depth texture
 - Fixed issue with shader graph custom velocity and VFX (case 1388149)
 - Fixed motion vector rendering with shader graph with planar primitive (case 1398313)
+- Fixed issue in APV with scenes saved only once when creating them.
+- Fixed probe volume baking not generating any probes on mac.
+- Fix a few UX issues in APV.
+- Fixed issue with detail normals when scale is null (case 1399548).
 
 ## [14.0.0] - 2021-11-17